refactor Realm constructor wireup
