+++ lib/index.js
@@ -42,44 +42,6 @@ if (typeof Realm != 'undefined') {
     throw new Error('Missing Realm constructor - please ensure RealmReact framework is included!');
 }
 
-// Add the specified Array methods to the Collection prototype.
-Object.defineProperties(realmConstructor.Collection.prototype, require('./collection-methods'));
-
-// Add sync methods
-if (realmConstructor.Sync) {
-    realmConstructor.Sync.User = require('./sync').User;
-}
-
-// TODO: Remove this now useless object.
-var types = Object.freeze({
-    'BOOL': 'bool',
-    'INT': 'int',
-    'FLOAT': 'float',
-    'DOUBLE': 'double',
-    'STRING': 'string',
-    'DATE': 'date',
-    'DATA': 'data',
-    'OBJECT': 'object',
-    'LIST': 'list',
-});
-Object.defineProperty(realmConstructor, 'Types', {
-    get: function() {
-        if (typeof console != 'undefined') {
-            /* global console */
-            /* eslint-disable no-console */
-            var stack = new Error().stack.split("\n").slice(2).join("\n");
-            var msg = '`Realm.Types` is deprecated! Please specify the type name as lowercase string instead!\n'+stack;
-            if (console.warn != undefined) {
-                console.warn(msg);
-            }
-            else {
-                console.log(msg);
-            }
-            /* eslint-enable no-console */
-        }
-        return types;
-    },
-    configurable: true
-});
+require('./wireup')(realmConstructor);
 
 module.exports = realmConstructor;
